add item CSS class & styles to BitMenuButton #4983
