+++ src/BlazorUI/Bit.BlazorUI/Components/Buttons/BitMenuButton/BitMenuButton.razor.cs
@@ -5,19 +5,23 @@ namespace Bit.BlazorUI;
 
 public partial class BitMenuButton<TItem> : IDisposable where TItem : class
 {
-    private const string IS_ENABLED_FIELD = nameof(BitMenuButtonItem.IsEnabled);
+    private const string CLASS_FIELD = nameof(BitMenuButtonItem.Class);
     private const string ICON_NAME_FIELD = nameof(BitMenuButtonItem.IconName);
-    private const string TEXT_FIELD = nameof(BitMenuButtonItem.Text);
+    private const string IS_ENABLED_FIELD = nameof(BitMenuButtonItem.IsEnabled);
     private const string KEY_FIELD = nameof(BitMenuButtonItem.Key);
+    private const string STYLE_FIELD = nameof(BitMenuButtonItem.Style);
+    private const string TEXT_FIELD = nameof(BitMenuButtonItem.Text);
 
 
     private BitButtonStyle buttonStyle = BitButtonStyle.Primary;
     private bool isCalloutOpen;
 
-    private string _internalIsEnabledField = IS_ENABLED_FIELD;
+    private string _internalClassField = CLASS_FIELD;
     private string _internalIconNameField = ICON_NAME_FIELD;
-    private string _internalTextField = TEXT_FIELD;
+    private string _internalIsEnabledField = IS_ENABLED_FIELD;
     private string _internalKeyField = KEY_FIELD;
+    private string _internalStyleField = STYLE_FIELD;
+    private string _internalTextField = TEXT_FIELD;
 
 
     private string? _calloutId;
@@ -88,6 +92,16 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
     [Parameter] public RenderFragment? ChildContent { get; set; }
 
     /// <summary>
+    /// The CSS Class field name of the custom input class.
+    /// </summary>
+    [Parameter] public string ClassField { get; set; } = CLASS_FIELD;
+
+    /// <summary>
+    /// The CSS Class field selector of the custom input class.
+    /// </summary>
+    [Parameter] public Expression<Func<TItem, bool>>? ClassFieldSelector { get; set; }
+
+    /// <summary>
     /// The content inside the header of MenuButton can be customized.
     /// </summary>
     [Parameter] public RenderFragment? HeaderTemplate { get; set; }
@@ -98,42 +112,42 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
     [Parameter] public string? IconName { get; set; }
 
     /// <summary>
+    /// IconName field name of the custom input class.
+    /// </summary>
+    [Parameter] public string IconNameField { get; set; } = ICON_NAME_FIELD;
+
+    /// <summary>
+    /// IconName field selector of the custom input class.
+    /// </summary>
+    [Parameter] public Expression<Func<TItem, string>>? IconNameFieldSelector { get; set; }
+
+    /// <summary>
+    /// IsEnabled field name of the custom input class.
+    /// </summary>
+    [Parameter] public string IsEnabledField { get; set; } = IS_ENABLED_FIELD;
+
+    /// <summary>
+    /// IsEnabled field selector of the custom input class.
+    /// </summary>
+    [Parameter] public Expression<Func<TItem, bool>>? IsEnabledFieldSelector { get; set; }
+
+    /// <summary>
     ///  List of BitMenuButtonItem to show as a item in MenuButton.
     /// </summary>
     [Parameter] public IEnumerable<TItem> Items { get; set; } = new List<TItem>();
 
     /// <summary>
-    /// Whether or not the item is enabled.
-    /// </summary>
-    [Parameter] public string IsEnabledField { get; set; } = IS_ENABLED_FIELD;
-
-    /// <summary>
-    /// Whether or not the item is enabled.
-    /// </summary>
-    [Parameter] public Expression<Func<TItem, bool>>? IsEnabledFieldSelector { get; set; }
-
-    /// <summary>
-    /// Name of an icon to render next to the item text.
-    /// </summary>
-    [Parameter] public string IconNameField { get; set; } = ICON_NAME_FIELD;
-
-    /// <summary>
-    /// Name of an icon to render next to the item text.
-    /// </summary>
-    [Parameter] public Expression<Func<TItem, string>>? IconNameFieldSelector { get; set; }
-
-    /// <summary>
-    /// The content inside the MenuButton-item can be customized.
+    /// The custom content to render each item.
     /// </summary>
     [Parameter] public RenderFragment<TItem>? ItemTemplate { get; set; }
 
     /// <summary>
-    /// A unique value to use as a key of the item.
+    /// Key field name of the custom input class.
     /// </summary>
     [Parameter] public string KeyField { get; set; } = KEY_FIELD;
 
     /// <summary>
-    /// A unique value to use as a key of the item.
+    /// Key field selector of the custom input class.
     /// </summary>
     [Parameter] public Expression<Func<TItem, string>>? KeyFieldSelector { get; set; }
 
@@ -148,17 +162,27 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
     [Parameter] public EventCallback<TItem> OnItemClick { get; set; }
 
     /// <summary>
+    /// The CSS Style field name of the custom input class.
+    /// </summary>
+    [Parameter] public string StyleField { get; set; } = STYLE_FIELD;
+
+    /// <summary>
+    /// The CSS Style field selector of the custom input class.
+    /// </summary>
+    [Parameter] public Expression<Func<TItem, string>>? StyleFieldSelector { get; set; }
+
+    /// <summary>
     /// The text to show inside the header of MenuButton.
     /// </summary>
     [Parameter] public string? Text { get; set; }
 
     /// <summary>
-    /// Name of an icon to render next to the item text.
+    /// Text field name of the custom input class.
     /// </summary>
     [Parameter] public string TextField { get; set; } = TEXT_FIELD;
 
     /// <summary>
-    /// Name of an icon to render next to the item text.
+    /// Text field selector of the custom input class.
     /// </summary>
     [Parameter] public Expression<Func<TItem, string>>? TextFieldSelector { get; set; }
 
@@ -191,10 +215,12 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
         _calloutId = $"{RootElementClass}-callout-{UniqueId}";
         _overlayId = $"{RootElementClass}-overlay-{UniqueId}";
 
+        _internalClassField = ClassFieldSelector?.GetName() ?? ClassField;
         _internalIsEnabledField = IsEnabledFieldSelector?.GetName() ?? IsEnabledField;
         _internalIconNameField = IconNameFieldSelector?.GetName() ?? IconNameField;
-        _internalTextField = TextFieldSelector?.GetName() ?? TextField;
         _internalKeyField = KeyFieldSelector?.GetName() ?? KeyField;
+        _internalStyleField = StyleFieldSelector?.GetName() ?? StyleField;
+        _internalTextField = TextFieldSelector?.GetName() ?? TextField;
 
         _dotnetObj = DotNetObjectReference.Create(this);
 
@@ -225,7 +251,20 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
         ClassBuilder.Register(() => _isCalloutOpen ? $"{RootElementClass}-omn" : string.Empty);
     }
 
+    private string? GetClass(TItem item)
+    {
+        if (item is BitMenuButtonItem bitMenuButtonItem)
+        {
+            return bitMenuButtonItem.Class;
+        }
 
+        if (item is BitMenuButtonOption menuButtonOption)
+        {
+            return menuButtonOption.Class;
+        }
+
+        return item.GetValueFromProperty<string?>(_internalClassField);
+    }
 
     private string? GetIconName(TItem item)
     {
@@ -240,36 +279,6 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
         }
 
         return item.GetValueFromProperty<string?>(_internalIconNameField);
-    }
-
-    private string? GetText(TItem item)
-    {
-        if (item is BitMenuButtonItem bitMenuButtonItem)
-        {
-            return bitMenuButtonItem.Text;
-        }
-
-        if (item is BitMenuButtonOption menuButtonOption)
-        {
-            return menuButtonOption.Text;
-        }
-
-        return item.GetValueFromProperty<string?>(_internalTextField);
-    }
-
-    private string? GetKey(TItem item)
-    {
-        if (item is BitMenuButtonItem bitMenuButtonItem)
-        {
-            return bitMenuButtonItem.Key;
-        }
-
-        if (item is BitMenuButtonOption menuButtonOption)
-        {
-            return menuButtonOption.Key;
-        }
-
-        return item.GetValueFromProperty<string?>(_internalKeyField);
     }
 
     private bool GetIsEnabled(TItem item)
@@ -287,6 +296,51 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
         return item.GetValueFromProperty(_internalIsEnabledField, true);
     }
 
+    private string? GetKey(TItem item)
+    {
+        if (item is BitMenuButtonItem bitMenuButtonItem)
+        {
+            return bitMenuButtonItem.Key;
+        }
+
+        if (item is BitMenuButtonOption menuButtonOption)
+        {
+            return menuButtonOption.Key;
+        }
+
+        return item.GetValueFromProperty<string?>(_internalKeyField);
+    }
+
+    private string? GetStyle(TItem item)
+    {
+        if (item is BitMenuButtonItem bitMenuButtonItem)
+        {
+            return bitMenuButtonItem.Style;
+        }
+
+        if (item is BitMenuButtonOption menuButtonOption)
+        {
+            return menuButtonOption.Style;
+        }
+
+        return item.GetValueFromProperty<string?>(_internalStyleField);
+    }
+
+    private string? GetText(TItem item)
+    {
+        if (item is BitMenuButtonItem bitMenuButtonItem)
+        {
+            return bitMenuButtonItem.Text;
+        }
+
+        if (item is BitMenuButtonOption menuButtonOption)
+        {
+            return menuButtonOption.Text;
+        }
+
+        return item.GetValueFromProperty<string?>(_internalTextField);
+    }
+
     private async Task OpenCallout(MouseEventArgs e)
     {
         if (IsEnabled is false) return;
@@ -297,7 +351,7 @@ public partial class BitMenuButton<TItem> : IDisposable where TItem : class
         await OnClick.InvokeAsync(e);
     }
 
-    internal async Task HandleOnItemClick(TItem item)
+    private async Task HandleOnItemClick(TItem item)
     {
         if (IsEnabled is false || GetIsEnabled(item) is false) return;
