+++ src/BlazorUI/Bit.BlazorUI/Components/Buttons/BitMenuButton/BitMenuButtonOption.cs
@@ -8,24 +8,34 @@ public class BitMenuButtonOption : ComponentBase, IDisposable
 
 
     /// <summary>
+    /// The custom CSS classes of the option.
+    /// </summary>
+    [Parameter] public string? Class { get; set; }
+
+    /// <summary>
+    /// Name of an icon to render next to the option text.
+    /// </summary>
+    [Parameter] public string? IconName { get; set; }
+
+    /// <summary>
     /// Whether or not the option is enabled.
     /// </summary>
     [Parameter] public bool IsEnabled { get; set; } = true;
 
     /// <summary>
-    /// Name of an icon to render next to the option text
-    /// </summary>
-    [Parameter] public string? IconName { get; set; }
-
-    /// <summary>
-    /// Text to render in the option
-    /// </summary>
-    [Parameter] public string? Text { get; set; }
-
-    /// <summary>
-    /// A unique value to use as a key of the option
+    /// A unique value to use as a key of the option.
     /// </summary>
     [Parameter] public string? Key { get; set; }
+
+    /// <summary>
+    /// The custom value for the style attribute of the option.
+    /// </summary>
+    [Parameter] public string? Style { get; set; }
+
+    /// <summary>
+    /// Text to render in the option.
+    /// </summary>
+    [Parameter] public string? Text { get; set; }
 
 
     protected override async Task OnInitializedAsync()
