--- src/BlazorUI/Demo/Client/Core/Pages/Components/Buttons/BitMenuButtonDemo.razor.scss
@@ -1,30 +1,0 @@
-﻿.example-content {
-    gap: 0.5rem;
-    display: flex;
-    flex-flow: row wrap;
-    align-items: center;
-    margin-bottom: 30px;
-
-    &.column {
-        flex-direction: column;
-        align-items: flex-start;
-    }
-}
-
-.clicked-item {
-    font-weight: 600;
-    margin-top: 0.5rem;
-}
-
-
-.item-template-box {
-    width: 100%;
-    display: flex;
-}
-
-::deep {
-    .custom-class {
-        color: aqua;
-        border-radius: 1rem;
-    }
-}
