+++ src/BlazorUI/Demo/Client/Core/Pages/Components/Buttons/MenuButton/MenuActionItem.cs
@@ -6,4 +6,7 @@ public class MenuActionItem
     public string? Name { get; set; }
     public string? Icon { get; set; }
     public bool IsEnabled { get; set; } = true;
+
+    public string? Class { get; set; }
+    public string? Style { get; set; }
 }
