+++ src/BlazorUI/Demo/Client/Core/Pages/Components/Buttons/MenuButton/_BitMenuButtonItemDemo.razor
@@ -1,0 +1,145 @@
+﻿<ComponentExampleBox Title="Basic" HTMLSourceCode="@example1BitMenuButtonItemHTMLCode" CSharpSourceCode="@example1BitMenuButtonItemCSharpCode" ExampleId="example11">
+    <ExamplePreview>
+        <div class="example-content">
+            <BitMenuButton Text="Primary"
+                           ButtonStyle="BitButtonStyle.Primary"
+                           Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example1SelectedItem = item.Key" />
+
+            <BitMenuButton Text="Standard"
+                           ButtonStyle="BitButtonStyle.Standard"
+                           Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example1SelectedItem = item.Key" />
+
+            <BitMenuButton Text="Disabled"
+                           Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example1SelectedItem = item.Key"
+                           IsEnabled="false" />
+
+            <BitMenuButton Text="Item Disabled"
+                           Items="itemDisabledItems"
+                           OnItemClick="(BitMenuButtonItem item) => example1SelectedItem = item.Key" />
+        </div>
+        <div class="clicked-item">Clicked Item: @example1SelectedItem</div>
+    </ExamplePreview>
+</ComponentExampleBox>
+
+<ComponentExampleBox Title="IconName" HTMLSourceCode="@example2BitMenuButtonItemHTMLCode" CSharpSourceCode="@example2BitMenuButtonItemCSharpCode" ExampleId="example12">
+    <ExamplePreview>
+        <div class="example-content">
+            <BitMenuButton Text="Primary Button"
+                           IconName="@BitIconName.Edit"
+                           ButtonStyle="BitButtonStyle.Primary"
+                           Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example2SelectedItem = item.Key" />
+
+            <BitMenuButton Text="Standard Button"
+                           IconName="@BitIconName.Add"
+                           ButtonStyle="BitButtonStyle.Standard"
+                           Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example2SelectedItem = item.Key" />
+        </div>
+        <div class="clicked-item">Clicked Item: @example2SelectedItem</div>
+    </ExamplePreview>
+</ComponentExampleBox>
+
+<ComponentExampleBox Title="Style & Class" HTMLSourceCode="@example3BitMenuButtonItemHTMLCode" CSharpSourceCode="@example3BitMenuButtonItemCSharpCode" ExampleId="example13">
+    <ExamplePreview>
+        <div>Component's Style & Class:</div>
+        <br />
+        <div class="example-content">
+            <BitMenuButton Text="Styled Button"
+                           Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example31SelectedItem = item.Key"
+                           Style="width: 200px; height: 40px; background-color: #8A8886; border-color: black;" />
+
+            <BitMenuButton Text="Classed Button"
+                           Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example31SelectedItem = item.Key"
+                           Class="custom-class" />
+        </div>
+        <div class="clicked-item">Clicked Item: @example31SelectedItem</div>
+        <br /><br /><br /><br />
+        <div>Item's Style & Class:</div>
+        <br />
+        <div class="example-content">
+            <BitMenuButton Text="Item Styled & Classed Button"
+                           Items="itemStyleClassItems"
+                           OnItemClick="(BitMenuButtonItem item) => example32SelectedItem = item.Key" />
+        </div>
+        <div class="clicked-item">Clicked Item: @example32SelectedItem</div>
+    </ExamplePreview>
+</ComponentExampleBox>
+
+<ComponentExampleBox Title="HeaderTemplate" HTMLSourceCode="@example4BitMenuButtonItemHTMLCode" CSharpSourceCode="@example4BitMenuButtonItemCSharpCode" ExampleId="example14">
+    <ExamplePreview>
+        <div class="example-content">
+            <BitMenuButton Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example4SelectedItem = item.Key"
+                           ButtonStyle="BitButtonStyle.Primary">
+                <HeaderTemplate>
+                    <BitIcon IconName="@BitIconName.Warning" />
+                    <div style="font-weight: 600; color: white;">
+                        Custom Header!
+                    </div>
+                    <BitIcon IconName="@BitIconName.Warning" />
+                </HeaderTemplate>
+            </BitMenuButton>
+
+            <BitMenuButton Items="basicItems"
+                           OnItemClick="(BitMenuButtonItem item) => example4SelectedItem = item.Key"
+                           ButtonStyle="BitButtonStyle.Standard">
+                <HeaderTemplate>
+                    <div style="font-weight: bold; color: #d13438;">
+                        Custom Header!
+                    </div>
+                </HeaderTemplate>
+            </BitMenuButton>
+        </div>
+        <div class="clicked-item">Clicked Item: @example4SelectedItem</div>
+    </ExamplePreview>
+</ComponentExampleBox>
+
+<ComponentExampleBox Title="ItemTemplate" HTMLSourceCode="@example5BitMenuButtonItemHTMLCode" CSharpSourceCode="@example5BitMenuButtonItemCSharpCode" ExampleId="example15">
+    <ExamplePreview>
+        <div class="example-content">
+            <BitMenuButton Text="Primary Button"
+                           IconName="@BitIconName.Edit"
+                           Items="itemTemplateItems"
+                           OnItemClick="(BitMenuButtonItem item) => example5SelectedItem = item.Key"
+                           ButtonStyle="BitButtonStyle.Primary">
+                <ItemTemplate Context="item">
+                    <div class="item-template-box">
+                        <span style="color: @(item.Key == "add-key" ? "green" : item.Key == "edit-key" ? "yellow" : "red");">
+                            @item.Text (@item.Key)
+                        </span>
+                    </div>
+                </ItemTemplate>
+            </BitMenuButton>
+
+            <BitMenuButton Text="Standard Button"
+                           IconName="@BitIconName.Edit"
+                           Items="itemTemplateItems"
+                           OnItemClick="(BitMenuButtonItem item) => example5SelectedItem = item.Key"
+                           ButtonStyle="BitButtonStyle.Standard">
+                <ItemTemplate Context="item">
+                    <div class="item-template-box">
+                        <span style="color: @(item.Key == "add-key" ? "green" : item.Key == "edit-key" ? "yellow" : "red");">
+                            @item.Text (@item.Key)
+                        </span>
+                    </div>
+                </ItemTemplate>
+            </BitMenuButton>
+        </div>
+        <div class="clicked-item">Clicked Item: @example5SelectedItem</div>
+    </ExamplePreview>
+</ComponentExampleBox>
+
+@code {
+    private string? example1SelectedItem;
+    private string? example2SelectedItem;
+    private string? example31SelectedItem;
+    private string? example32SelectedItem;
+    private string? example4SelectedItem;
+    private string? example5SelectedItem;
+}
